unslopped the AI output
went through the code with a laundry list of things to tighten up
